[Fix] 슈팅 게임 버그 수정중
현재 버그 리스트

1.카메라 관련 연출 버그
2.카드 자동 선택 관련 버그
