+++ Assets/PMS/PMS_Scripts/FollowCam.cs
@@ -1,0 +1,20 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class FollowCam : MonoBehaviour
+{
+    public Transform target;
+
+    private void LateUpdate()
+    {
+ 
+    }
+
+    private void Update()
+    {
+        //if(target == null) return;
+        //Vector3 newPos = new Vector3(target.position.x, transform.position.y,target.position.z);
+        //transform.position = newPos;
+    }
+}
