--- Assets/PMS/PMS_Scripts/FollowTarget.cs
@@ -1,17 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class FollowTarget : MonoBehaviour
-{
-    public Transform target;
-    public Vector3 offset = new Vector3(0, 15, 0);
-
-
-    private void LateUpdate()
-    {
-        if (target == null) return;
-        transform.position = target.position + offset;
-        transform.LookAt(target);
-    }
-}
