+++ Assets/PMS/PMS_Scripts/GameState/ShootingGameState/CardSelectState.cs
@@ -17,8 +17,8 @@ public class CardSelectState : ShootingGameState
             double startAt = PhotonNetwork.Time + lead;
             double endAt = startAt + duration;
             CardManager.Instance.BuildAndBroadcastDeck();
-            ShootingNetworkManager.Instance.photonView.RPC("RPC_StartTimer", RpcTarget.All, startAt,endAt);
-            CardManager.Instance.StartAutoCardSelect();
+            //ShootingNetworkManager.Instance.photonView.RPC("RPC_StartTimer", RpcTarget.All, startAt,endAt);
+            //CardManager.Instance.StartAutoCardSelect();
         }
         else
         {
