+++ Assets/Photon/PhotonUnityNetworking/Resources/PhotonServerSettings.asset
@@ -119,6 +119,7 @@ MonoBehaviour:
   - RPC_Stun
   - RPC_SyncRanks
   - SetOverheat
+  - UnEnableInteraction
   DisableAutoOpenWizard: 1
   ShowSettings: 0
   DevRegionSetOnce: 1
